Restructure packages for online banking core
COAND-497
