+++ online-banking-core/src/main/java/com/adyen/checkout/onlinebankingcore/OnlineBankingComponent.kt
@@ -28,6 +28,7 @@ import com.adyen.checkout.components.ui.ViewableComponent
 import com.adyen.checkout.components.ui.view.ComponentViewType
 import com.adyen.checkout.core.log.LogUtil
 import com.adyen.checkout.core.log.Logger
+import com.adyen.checkout.onlinebankingcore.internal.ui.OnlineBankingDelegate
 import kotlinx.coroutines.flow.Flow
 
 abstract class OnlineBankingComponent<IssuerListPaymentMethodT : IssuerListPaymentMethod> protected constructor(
