+++ online-banking-core/src/main/java/com/adyen/checkout/onlinebankingcore/internal/ui/OnlineBankingDelegate.kt
@@ -6,9 +6,10 @@
  * Created by atef on 8/9/2022.
  */
 
-package com.adyen.checkout.onlinebankingcore
+package com.adyen.checkout.onlinebankingcore.internal.ui
 
 import android.content.Context
+import androidx.annotation.RestrictTo
 import com.adyen.checkout.components.PaymentComponentState
 import com.adyen.checkout.components.base.PaymentComponentDelegate
 import com.adyen.checkout.components.model.payments.request.IssuerListPaymentMethod
@@ -16,8 +17,12 @@ import com.adyen.checkout.components.ui.ButtonDelegate
 import com.adyen.checkout.components.ui.UIStateDelegate
 import com.adyen.checkout.components.ui.ViewProvidingDelegate
 import com.adyen.checkout.core.exception.CheckoutException
+import com.adyen.checkout.onlinebankingcore.internal.ui.model.OnlineBankingInputData
+import com.adyen.checkout.onlinebankingcore.internal.ui.model.OnlineBankingModel
+import com.adyen.checkout.onlinebankingcore.internal.ui.model.OnlineBankingOutputData
 import kotlinx.coroutines.flow.Flow
 
+@RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
 interface OnlineBankingDelegate<IssuerListPaymentMethodT : IssuerListPaymentMethod> :
     PaymentComponentDelegate<PaymentComponentState<IssuerListPaymentMethodT>>,
     ViewProvidingDelegate,
