+++ online-banking-core/src/main/java/com/adyen/checkout/onlinebankingcore/internal/ui/model/OnlineBankingInputData.kt
@@ -6,10 +6,12 @@
  * Created by atef on 20/9/2022.
  */
 
-package com.adyen.checkout.onlinebankingcore
+package com.adyen.checkout.onlinebankingcore.internal.ui.model
 
+import androidx.annotation.RestrictTo
 import com.adyen.checkout.components.base.InputData
 
+@RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
 data class OnlineBankingInputData(
     var selectedIssuer: OnlineBankingModel? = null,
 ) : InputData
