+++ online-banking-core/src/main/java/com/adyen/checkout/onlinebankingcore/internal/ui/model/OnlineBankingOutputData.kt
@@ -6,12 +6,15 @@
  * Created by atef on 20/9/2022.
  */
 
-package com.adyen.checkout.onlinebankingcore
+package com.adyen.checkout.onlinebankingcore.internal.ui.model
 
+import androidx.annotation.RestrictTo
 import com.adyen.checkout.components.base.OutputData
 import com.adyen.checkout.components.ui.FieldState
 import com.adyen.checkout.components.ui.Validation
+import com.adyen.checkout.onlinebankingcore.R
 
+@RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
 data class OnlineBankingOutputData(
     val selectedIssuer: OnlineBankingModel? = null
 ) : OutputData {
