+++ online-banking-core/src/test/java/com/adyen/checkout/onlinebankingcore/internal/ui/DefaultOnlineBankingDelegateTest.kt
@@ -1,12 +1,12 @@
 /*
- * Copyright (c) 2022 Adyen N.V.
+ * Copyright (c) 2023 Adyen N.V.
  *
  * This file is open source and available under the MIT license. See the LICENSE file for more info.
  *
- * Created by atef on 20/9/2022.
+ * Created by atef on 15/2/2023.
  */
 
-package com.adyen.checkout.onlinebankingcore
+package com.adyen.checkout.onlinebankingcore.internal.ui
 
 import android.content.Context
 import app.cash.turbine.test
@@ -19,6 +19,9 @@ import com.adyen.checkout.components.model.payments.request.OrderRequest
 import com.adyen.checkout.components.repository.PaymentObserverRepository
 import com.adyen.checkout.components.ui.SubmitHandler
 import com.adyen.checkout.core.api.Environment
+import com.adyen.checkout.onlinebankingcore.internal.ui.model.OnlineBankingModel
+import com.adyen.checkout.onlinebankingcore.internal.ui.model.OnlineBankingOutputData
+import com.adyen.checkout.onlinebankingcore.internal.util.PdfOpener
 import com.adyen.checkout.onlinebankingcore.utils.TestOnlineBankingConfiguration
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.ExperimentalCoroutinesApi
